Create passthrough_hp_common
